+++ src/main/java/server/drawwang/domain/board/service/implementation/event/listener/SubmittedBoardEventListener.java
@@ -1,4 +1,4 @@
-package server.drawwang.domain.board.service.implementation.event.Listener;
+package server.drawwang.domain.board.service.implementation.event.listener;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -6,29 +6,22 @@ import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.event.TransactionalEventListener;
 import server.drawwang.domain.board.entity.BoardEntity;
-import server.drawwang.domain.board.repository.BoardRepository;
-import server.drawwang.domain.board.service.implementation.event.BoardLikedEvent;
+import server.drawwang.domain.board.service.implementation.event.SubmittedBoardEvent;
 import server.drawwang.domain.thread.entity.ThreadEntity;
 import server.drawwang.domain.thread.repository.ThreadRepository;
-import server.drawwang.global.exception.CustomErrorCode;
-import server.drawwang.global.exception.CustomException;
 
 @Component
 @RequiredArgsConstructor
-public class BoardLikedEventListener {
+public class SubmittedBoardEventListener {
     private final ThreadRepository threadRepository;
-    private final BoardRepository boardRepository;
 
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     @TransactionalEventListener
-    public void updateThreadKing(BoardLikedEvent event) {
+    public void updateKingIfEmptyThread(SubmittedBoardEvent event) {
+        ThreadEntity threadEntity = event.getThreadEntity();
         BoardEntity boardEntity = event.getBoardEntity();
-        ThreadEntity threadEntity = boardEntity.getThread();
 
-        BoardEntity kingBoardEntity = boardRepository.findById(boardEntity.getThread().getKingBoardId())
-                .orElseThrow(() -> new CustomException(CustomErrorCode.THREAD_KING_NOT_FOUND_ERROR));
-
-        if (boardEntity.getLikes() > kingBoardEntity.getLikes()) {
+        if(threadEntity.getKingBoardId() == null) {
             threadEntity.setKingBoardId(boardEntity.getId());
             threadRepository.save(threadEntity);
         }
